rename Util folder to Utils, move DataError types to the utils folder
