+++ app/src/main/java/com/revakovskyi/nytbooks/di/AppModule.kt
@@ -1,7 +1,7 @@
 package com.revakovskyi.nytbooks.di
 
-import com.revakovskyi.core.domain.util.DispatcherProvider
-import com.revakovskyi.core.domain.util.StandardDispatchers
+import com.revakovskyi.core.domain.utils.DispatcherProvider
+import com.revakovskyi.core.domain.utils.StandardDispatchers
 import com.revakovskyi.nytbooks.presentation.MainViewModel
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.core.module.dsl.singleOf
