--- core/domain/src/main/java/com/revakovskyi/core/domain/util/Error.kt
@@ -1,3 +1,0 @@
-package com.revakovskyi.core.domain.util
-
-interface Error
