+++ core/domain/src/main/java/com/revakovskyi/core/domain/utils/Error.kt
@@ -1,0 +1,3 @@
+package com.revakovskyi.core.domain.utils
+
+interface Error
