+++ core/domain/src/main/java/com/revakovskyi/core/domain/utils/Result.kt
@@ -1,9 +1,9 @@
-package com.revakovskyi.core.domain.util
+package com.revakovskyi.core.domain.utils
 
 sealed interface Result<out D, out E : Error> {
 
     data class Success<out D>(val data: D) : Result<D, Nothing>
-    data class Error<out E : com.revakovskyi.core.domain.util.Error>(val error: E) :
+    data class Error<out E : com.revakovskyi.core.domain.utils.Error>(val error: E) :
         Result<Nothing, E>
 
 }
