+++ core/network/src/main/java/com/revakovskyi/core/network/tools/HttpClientExt.kt
@@ -1,8 +1,8 @@
 package com.revakovskyi.core.network.tools
 
 import android.util.Log
-import com.revakovskyi.core.domain.books.DataError
-import com.revakovskyi.core.domain.util.Result
+import com.revakovskyi.core.domain.utils.DataError
+import com.revakovskyi.core.domain.utils.Result
 import com.revakovskyi.core.network.BuildConfig
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
